requirements.txt
anthropic
Pillow
keyboard
pystray
customtkinter
duckduckgo-search
pyautogui
python-dotenv
